gst/level/gstlevel.*: Fix type mixup in level->interval (gdouble<->guint64). Spotted by
Original commit message from CVS:
* gst/level/gstlevel.c: (gst_level_set_property):
* gst/level/gstlevel.h:
Fix type mixup in level->interval (gdouble<->guint64). Spotted by
René Stadler

+++ gst/level/gstlevel.h
@@ -53,7 +53,7 @@ struct _GstLevel {
   GstBaseTransform element;
 
   gboolean message;             /* whether or not to post messages */
-  gdouble interval;             /* how many seconds between emits */
+  guint64 interval;             /* how many seconds between emits */
 
   gint rate;                    /* caps variables */
   gint width;
